More changes to activate force training.
In this commit, forces can be obtained with autograd with:

```
    forces = -torch.autograd.grad(outputs[0], example, create_graph=True, retain_graph=True)[0]
```

This is still to be tested during training.

+++ ml4chem/atomistic/models/base.py
@@ -65,7 +65,7 @@ class DeepLearningModel(ABC):
                     for symbol, features_ in values:
                         if symbol not in image.keys():
                             image[symbol] = []
-                        image[symbol].append(features_)
+                        image[symbol].append(features_.float())
 
                     for symbol in data.unique_element_symbols[purpose]:
                         tensors = image.get(symbol)
